requirements.txt
django-allauth==0.60.1